# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/wms-rsd/barang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Daftar barang dengan pagination \u0026 filter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Daftar barang",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Item per halaman",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kata kunci pencarian",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter kategori",
                        "name": "kategori_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter satuan",
                        "name": "satuan_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Hanya tampilkan stok menipis",
                        "name": "stok_menipis",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "aktif untuk hanya barang aktif",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Tambah barang baru",
                "parameters": [
                    {
                        "description": "Data barang",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/barang.BarangRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Envelope"
                        }
                    },
                    "409": {
                        "description": "kode barang sudah dipakai",
                        "schema": {
                            "$ref": "#/definitions/utils.Envelope"
                        }
                    }
                }
            }
        },
        "/wms-rsd/barang/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Detail barang",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID barang",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Envelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "barang.BarangRequest": {
            "type": "object",
            "required": [
                "kategori_id",
                "kode_barang",
                "nama",
                "satuan_id"
            ],
            "properties": {
                "deskripsi": {
                    "type": "string",
                    "maxLength": 255
                },
                "harga_beli": {
                    "type": "integer",
                    "minimum": 0
                },
                "kategori_id": {
                    "type": "integer"
                },
                "kode_barang": {
                    "type": "string",
                    "maxLength": 30
                },
                "nama": {
                    "type": "string",
                    "maxLength": 150
                },
                "satuan_id": {
                    "type": "integer"
                },
                "stok_minimum": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "utils.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}